fixed narrowing conversion warning

+++ rplidarsdk/rplidar_driver.cpp
@@ -1790,7 +1790,7 @@ u_result RPlidarDriverImplCommon::grabScanData(rplidar_response_measurement_node
 {
     DEPRECATED_WARN("grabScanData()", "grabScanDataHq()");
 
-    switch (_dataEvt.wait(timeout))
+    switch ((int)_dataEvt.wait(timeout))
     {
     case rp::hal::Event::EVENT_TIMEOUT:
         count = 0;
@@ -1819,7 +1819,7 @@ u_result RPlidarDriverImplCommon::grabScanData(rplidar_response_measurement_node
 
 u_result RPlidarDriverImplCommon::grabScanDataHq(rplidar_response_measurement_node_hq_t * nodebuffer, size_t & count, _u32 timeout)
 {
-    switch (_dataEvt.wait(timeout))
+    switch ((int)_dataEvt.wait(timeout))
     {
     case rp::hal::Event::EVENT_TIMEOUT:
         count = 0;
